update project structure for more readability
