+++ plugins/out_oracle_log_analytics/CMakeLists.txt
@@ -1,6 +1,7 @@
 set(src
    oci_logan.c
    oci_logan_conf.c
+   oci_logan_helper.c
    )
 
 FLB_PLUGIN(out_oracle_log_analytics "${src}" "")
